📝 Refactor: Swagger 문서 수정 (#28)
