+++ src/main/java/com/skhu/moodfriend/app/controller/ProfileController.java
@@ -1,5 +1,6 @@
 package com.skhu.moodfriend.app.controller;
 
+import io.swagger.v3.oas.annotations.Hidden;
 import org.springframework.core.env.Environment;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -7,6 +8,7 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.Arrays;
 import java.util.List;
 
+@Hidden
 @RestController
 public class ProfileController {
