+++ src/main/java/com/skhu/moodfriend/app/controller/chat/HoyaController.java
@@ -15,7 +15,7 @@ import java.security.Principal;
 
 @RestController
 @RequiredArgsConstructor(access = AccessLevel.PROTECTED)
-@Tag(name = "호야 챗봇", description = "호야 챗봇 api 그룹")
+@Tag(name = "호야 챗봇", description = "호야 챗봇을 담당하는 api 그룹")
 @RequestMapping("/api/v1/hoya/chat")
 public class HoyaController {
 
@@ -23,10 +23,10 @@ public class HoyaController {
 
     @PostMapping
     @Operation(
-            summary = "챗봇 응답 조회",
-            description = "사용자의 프롬프트에 대한 챗봇 응답을 조회합니다.",
+            summary = "챗봇 응답",
+            description = "사용자의 프롬프트에 대한 챗봇 응답을 받습니다.",
             responses = {
-                    @ApiResponse(responseCode = "200", description = "챗봇 응답 조회 성공"),
+                    @ApiResponse(responseCode = "200", description = "챗봇 응답 성공"),
                     @ApiResponse(responseCode = "400", description = "잘못된 요청 or 입력값 오류"),
                     @ApiResponse(responseCode = "403", description = "권한 문제 or 관리자 문의"),
                     @ApiResponse(responseCode = "500", description = "서버 문제 or 관리자 문의")
