+++ src/main/java/com/skhu/moodfriend/app/controller/tracker/DiaryAIController.java
@@ -1,5 +1,7 @@
 package com.skhu.moodfriend.app.controller.tracker;
 
+import com.skhu.moodfriend.app.dto.tracker.resDto.DiaryAIResDto;
+import com.skhu.moodfriend.app.service.tracker.DiaryAIDisplayService;
 import com.skhu.moodfriend.app.service.tracker.DiaryAIModifyService;
 import com.skhu.moodfriend.global.template.ApiResponseTemplate;
 import io.swagger.v3.oas.annotations.Operation;
@@ -10,13 +12,16 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 @RestController
 @RequiredArgsConstructor(access = AccessLevel.PROTECTED)
-@Tag(name = "AI 일기 삭제", description = "AI 일기를 삭제하는 api 그룹")
+@Tag(name = "AI 일기", description = "AI 일기를 담당하는 api 그룹")
 @RequestMapping("/api/v1/diary/ai")
 public class DiaryAIController {
 
     private final DiaryAIModifyService diaryAIModifyService;
+    private final DiaryAIDisplayService diaryAIDisplayService;
 
     @DeleteMapping("delete/{diaryAIId}")
     @Operation(
@@ -34,6 +39,41 @@ public class DiaryAIController {
             @RequestParam Long memberId) {
 
         ApiResponseTemplate<Void> data = diaryAIModifyService.deleteDiaryAI(diaryAIId, memberId);
+        return ResponseEntity.status(data.getStatus()).body(data);
+    }
+
+    @GetMapping("/display/{diaryAIId}")
+    @Operation(
+            summary = "사용자의 특정 AI 일기 조회",
+            description = "사용자의 특정 AI 일기를 조회합니다.",
+            responses = {
+                    @ApiResponse(responseCode = "200", description = "특정 AI 일기 조회 성공"),
+                    @ApiResponse(responseCode = "403", description = "권한 문제 or 관리자 문의"),
+                    @ApiResponse(responseCode = "404", description = "해당 AI 일기를 찾을 수 없음"),
+                    @ApiResponse(responseCode = "500", description = "서버 문제 or 관리자 문의")
+            }
+    )
+    public ResponseEntity<ApiResponseTemplate<DiaryAIResDto>> getDiarySummaryById(
+            @PathVariable Long diaryAIId,
+            @RequestParam Long memberId) {
+
+        ApiResponseTemplate<DiaryAIResDto> data = diaryAIDisplayService.getDiarySummaryById(diaryAIId, memberId);
+        return ResponseEntity.status(data.getStatus()).body(data);
+    }
+
+    @GetMapping("/display")
+    @Operation(
+            summary = "사용자의 모든 AI 일기 조회",
+            description = "사용자의 모든 AI 일기를 조회합니다.",
+            responses = {
+                    @ApiResponse(responseCode = "200", description = "모든 AI 일기 조회 성공"),
+                    @ApiResponse(responseCode = "403", description = "권한 문제 or 관리자 문의"),
+                    @ApiResponse(responseCode = "500", description = "서버 문제 or 관리자 문의")
+            }
+    )
+    public ResponseEntity<ApiResponseTemplate<List<DiaryAIResDto>>> getAllDiaryAISummariesByUser(@RequestParam Long memberId) {
+
+        ApiResponseTemplate<List<DiaryAIResDto>> data = diaryAIDisplayService.getAllDiarySummariesByMember(memberId);
         return ResponseEntity.status(data.getStatus()).body(data);
     }
 }
